modificamos la libreria del senssor de luz para dar los datos en lux

+++ K32L2B31A_Project/iot_sdk/iot_sdk_peripheral/sensor_de_luz.h
@@ -42,7 +42,7 @@
 
 void sensorDeLuzIniciarCaptura(void);
  void sensorDeLuzEsperarResultado(void);
- uint32_t sensorDeLuzObtenerDatosADC(void);
+ float sensorDeLuzObtenerDatosADC(void);
 
 /** @} */ // end of X group
 /** @} */ // end of X group
